[semantic-arc] Add the ability to at the command line run a subset of semantic-arc optimizations.
SemanticARCOpts for maintainance reasons is becoming a series of small
optimization passes with a little pass manager that runs them. Given that, I
want to be able to be able to run subsets of these transforms for testing
reasons while normally running them in series.

This is in preparation for landing the coroutine lifetime extender and the
auto-coroutinizer.