Remove “folding” in SILBuilder APIs - emitReleaseValue/emitStrongRelease/emitDestroyValue
These functions did not just create the SILInstruction, they also look
for a retain and delete it when safe to do so. This can cause a crash in
SILCombine of partial_applys. SILCombine maintains a worklist of
instructions and deleting of instructions is allowed only via callbacks
which will also delete the instruction from the worklist.  In SILCombine
of partial_apply we call  swift::tryDeleteDeadClosure which in turn
calls SILBuilder::emitDestroyValueOperation which can potentially delete
a retain instruction leaving a stale entry in the worklist.
This PR cleans up all such API’s in SILBuilder, so that we don’t end up
with such bugs.

Not changing SILBuilder::emitDestroyAddrAndFold because it does not
delete any instructions, and tries to fold a destroy_addr with a
copy_addr converting it to a copy_addr [take].