Fix Mem2Reg for store [assign]
This change fixes the handling of store [assign] when a running value is
available in the current block for a multi-block use alloc_stack
instruction.

Discussion:
These are the following possibilities with a store [assign]

1. store [assign] in the entry block
   Running value is always available in this case, and we can create a
destroy_value of the running value.

2. store [assign] in a non-entry block with running value available
   This happens when there is another store to the stack location. And
this is not the first store in the block. Insert a destroy_value of the
running value

3. store [assign] in a non-entry block with no running value
   This happens when the store is the first in the block, then we let
fixBranchesAndUses to insert the destroy_value from the live out of
predecessor blocks.