[AST] Add simple LifetimeAnnotation enum.
The enum just provides a convenience around mutually exclusive
attributes which may be applied to decls and to SILFunctionArguments.
This way we can switch over the flavor of annotation that a decl enjoys
rather than having to remember the list of which might affect it, a list
which is likely to grow.