Update tests for allocbox to stack error change.
I also extended the tests to handle more interesting cases.

NOTE: There are a few cases where we introduced some new do not understand
errors. I am going to fix that in the next commit. I just wanted to completely
update the tests for the manner in which the allocbox to stack change affects
them.