[move-only] Add a test that shows the previous commit allows consume on var without errors to work.
I found that in the given case, an extra allocation was being emitted causing us
to in the success case to have an extra copy. In the previous commit, as part of
updating the codegen for consume, I also ensured that we wouldn't have that
extra allocation any more by handling the lvalue directly.

rdar://109222496