Fix mem2reg of lexical enum stack locations
This change avoids creating move_value [lexical] for trivial stores.

This can happen when we have a non trivial enum lexical stack location,
and we store a none value trivially.

Do the same when we have store_borrow of trivial values, even though there
is no ownership error in this case, handle it like store for consistency