Add isolation info to ExtInfo for SubscriptDecls
Subscripts with `isolated` arguments currently do not pass a compiler
assertion about the consistency of the function type w.r.t. isolation.
This is because the `ExtInfo` for `SubscriptDecl`s doesn't currently
have isolation information applied (though this works in practice).

Fix this by preserving the ExtInfo from the original interface type
of the subscript declaration, which is already what we do for
functions.